Add icon to account authenticator.
While the account itself is hidden from the accounts menu in Android it does
actually show up when trying to add a new account using Android settings. Since
there is no icon or title the entry is empty (but can be used). This looks ugly
and can be confusing.

Add title / icon to the account so it identifies itself correctly.