Attempt to fix preference style on Samsung devices.
On Samsung S7 the toolbar on preference screens was white. And icons in it were also white. This was because Car Report used the device default styles for preferences.

This fix uses the AppCompat theme for preferences now. This should make them look exactly like the rest of the app.